+++ example-frontend/store/syncDbSdk.ts
@@ -56,4 +56,7 @@ export const {
   useCreateMutation: useCreateTodo,
   useUpdateMutation: useUpdateTodo,
   useDeleteMutation: useDeleteTodo,
-} = createCollectionHooks<Todo, CreateTodoBody, UpdateTodoBody>({collection: "todos"});
+} = createCollectionHooks<Todo, CreateTodoBody, UpdateTodoBody>({
+  collection: "todos",
+  retries: false,
+});